Treat 0 the same as -1 and NULL for numeric columns such as "disc".
Update issue #411
Status: Fixed
Ah ha!  I knew it was the disc tag - one of them had a value of 0 set, and the other had no value at all which was taken by clementine to mean "-1", which was lower than 0 so screwed up the sorting.  Try now and it should be fixed (you'll have to get Clementine to rescan that directory - probably remove it from the library settings and add it again).