Save the saved streams after editing one.
Update issue #404
Status: Fixed
Ah I see the problem - it didn't save the streams to the config file when you edited one - only when you added a new one.  So if you edited a stream *then* added a new one it would save both, but if you edit one and then exit it would get lost.  Should be fixed now.