Don't replace the whole selection when autoscrolling.
Update issue #496
Status: Fixed
I think the autoscroll is useful even when the selected song isn't the
currently playing song.  I've changed it slightly now though so it won't
replace the selection when it scrolls, so you'll be able to press up or
down to go back to where you were before.