Style the Rip cd dialog
Set a specific tab order in the dialog.
Add a Clementine window icon.
Add a header for the first column in the table widget.
Change the button box to a cancel and a close button and make these behave as in the transcoder dialog.
Move the rip button to the button box.
Hide the progress bar if no cd has been ripped.