Fix oversized album cover art
When viewing the "Full Size" album covers, if the cover was higher in
resolution than that of the monitor, it would get cut off. Also,
the window was resizable however the image was not. This patch makes
the dialog a fixed size and resizes the cover only if it is too large to
fit on the screen. It also now displays the dimentions of the original
artwork in the title bar.