Add new now playing widget mode
In an attempt to satisfy many users, I have added another mode to the
now playing widget. This shows the large cover art however it draws the song details
below the artwork, without the gradient that everyone seems to have a problem with.
This is just an option and all the original functionality is unchanged.