Add ability to fit now playing cover to width
Some people seem to want to have a way of enlarging the now playing
widget cover when in large mode to get rid of the black bars on the sides.
This patch adds an option in the context menu which allows the cover to
fill the width of the panel when either large mode is enabled.