Fix Last.fm scrobbling after seek Fixes #4836 Last.fm defines a scrobble should be sent if: -the track is longer than 30 seconds. -the track has been played for at least half its duration, or for 4 minutes (whichever occurs earlier.)
Clementine has treated this as seconds from the start of the track, and if any seeking occurs, it nullifies the scrobble.
This IMO is incorrect. If I skip the first 10 seconds of a song, but listen to the rest (still meeting the time requirements),
I should still be able to scrobble the play. This change moves the scrobble point with every seek, requiring continuous playback
from any point that satisfies the time criteria.