Fix autoscroll fail to follow the selection in "InternetView"
The cause of this issue was the missing call of the parent class
method in "InternetView::currentChanged".

Removed "InternetView::CurrentIndexChanged" signal along with
the "InternetView::currentChanged" method since it looks redundant,
the "InternetView::selectionModel" and its "currentChanged" signal
can be used instead, also it isn't used anywhere in the code.

Resolves: #4485