Fix misbehavior when you double click after a single click on a tree item
This issue appeared as a side effect since the commit
bcaa9e4a37b732f91a36539e313937537be7aba2, which adds dropdown
animation to the library/internet view.

Reproduce:
Single/double click on a song item in library/internet view
then all double clicks on this item will be ignored, until you
single click on it once again.

The reason of this behavior in that the "setExpanded" method call from the
"ItemClicked" slot makes the invocation of the "mouseDoubleClickEvent"
method in the context where the view is in the "QAbstractTreeView::AnimatingState"
which makes impossible to emit "doubleClicked" signal, because it just
returns immediatelly (see http://code.qt.io/cgit/qt/qt.git/tree/src/gui/itemviews/qtreeview.cpp#n1849).

To get rid of this behavior we emit "doubleClicked" signal ourselves.