Bugfix for resume playback on start
This fixes issue #5365.

Because of the asynchronous loading of playlists introduced by
09e83935, the resume playback on startup logic finds the active
playlist as empty, because it is not restored yet. By attaching
the playback resume to the RestoreFinish signal, the playback
will be triggered as soon as the playlist is restored. It may be
possible (but unlikely) that the playlist will already be restored
before we wait for the signal, and playback won't be resumed.

Signed-off-by: Marko Hauptvogel <marko.hauptvogel@googlemail.com>