Fix C:\data cache directory for portable builds
Lastfm needs a NetworkAccessManager which is created during the
initialization of lastfm in main.cpp, it also sets a cache directory.

For the portable version, GetConfigPath uses
QCoreApplication::applicationDirPath() to get the current directory of
Clementine. If no QApplication (or QtSingleApplication in Clementines
case) was instantiated, it returns /data on Linux or C:\data on Windows
for the cache config path.

This commit changes the order so lastfm will be initialized after
QtSingleApplication was created and the correct cache config path will be
returned.

Fixes #5354