Revert "Blacklist all NVidia drivers"
In 2011, there was a bug that caused NVIDIA drivers to hang
Clementine on shutdown. In 2012, only some drivers had the
fix for this issue. Now, in 2019, we do not need to work
around this bug. By reverting commit

c2723008a22c5a4c4e307ba9a278fd2162ef6bdb

we work around known bad drivers, but do not penalize all
NVIDIA users for this ancient bug.