Fix rpmbuild error in librojectm
When using rpmbuild on Fedora 33, the build fails with:
error: ‘__builtin_strncpy’ specified bound depends on the length of the source argument [-Werror=stringop-overflow=]

This was fixed as part of projectm change d4d74f6d. This change takes the
relevant parts of that commit.

Note that the projectm commit contains several other warning fixes that might
affect Clementine builds in the future.