build: remove bad double negative when discovering 'git'
This automatically works because CMake's `if` statement considers
variables which' content ends with "-NOTFOUND" to be FALSE.

The previous version however was first tried as a variable which didn't
exist. As a result after that it's interpreted as a string (not a
variable anymore). So it skips the "-NOTFOUND" check and only does a
"string not empty" check. Which, as this is a string constant, is always
true.